spliting server up
moving some functionality into a seperate lib files, should be more to follow.
